# Replace tsc/babel with swc