--- a/jest.config.cjs
+++ b/jest.config.cjs
@@ -9,8 +9,7 @@ module.exports = {
     "**/?(*.)+(spec|test).+(ts|tsx|js)"
   ],
   "transform": {
-    "^.+\\.(ts|tsx)$": "ts-jest",
-    "^.+\\.(js)$": "babel-jest",
+    '^.+\\.(t|j)sx?$': '@swc/jest',
   },
   testEnvironment: "jsdom",
   transformIgnorePatterns: [