Our jest testing, and transpilation in general is performed with Babel and tsc. This PR replaces them with SWC for a significantly speed up to that part of our toolchain.

On my machine:
- Tanspilation of src and bench folder: 14 s => 3 s,
- Unit tests: ~60 s => ~12 s

The reason it's faster it to be found in the underlying architechture where SWC is written in Rust with more degrees of freedom, and TSC/Babel are js/ts projects

A bit of background on this topic
https://blog.logrocket.com/why-you-should-use-swc/
https://engineering.telia.no/blog/we-exchanged-babel-tsc-for-swc-and-it-was-worth-it