@HarelM 2022-02-02
Npm trends shows that swc is far from being as popular as babel.
I tend to wait until it gains more popularity as I fear there will be issues related to typescript due to its immaturity...