@birkskyum 2022-02-02
It's behind the scenes the foundation of parcel 2 and next 12, so with that in mind, the adoption is actually quite fast https://www.npmtrends.com/@swc/core-vs-next-vs-parcel