@HarelM 2022-02-02
@nyurik you seemed excited about this in the slack channel. Feel free to weigh in here... :-)
I think this is a small change and easy to revert, but I would hate if someone would waste a lot of time trying to understand why something is not working as expected and find out that the cause is this tool and that we adopted it too early.