@nyurik 2022-02-02
@HarelM you do have a point, and it is a bit tricky to test these kinds of changes. I would recommend releasing this change as its own minor (not patch) change, and if possible, wait for a bit between releasing last version and the next (and don't merge anything else in that time window).