@wipfli 2022-02-03
The motivation for this pull request seems to be faster compilation, i.e., better developer experience. However, if we cannot merge anything to main for a while, this would have a negative impact on how we work.