@HarelM 2022-02-03
I agree, not being able to merge for a while is not a good option.
Let's try it out, we could always revert back if we feel some things are not working as expected.
Currently, this only affects the tests which for me is an enabler to running the tests locally.
We should remember though, add a development team, that we should be suspicious of this change and flag things that we see that are causing issues, if any.