@birkskyum 2022-02-03
Thanks for all the thoughts. I sense a consensus on carefully moving forward with this, and allowing other merges as well, while being extra attentive to issues that in any way could be related to this change.