@wipfli 2022-02-03
Sounds good to me. Is the `eval()` really needed?