@birkskyum 2022-02-03
It seemed to me that try {eval} catch{} was a simpler more direct way to test "evaluates without errors" without having to use async/await, dynamic imports etc.