@wipfli 2022-02-03
If the change to `eval` was not needed for swc, then it would have been better placed in a separate pull request. But never mind, I was just curious...