@wipfli 2022-02-03
Did you look into https://github.com/swc-project/swc/issues/140?