@birkskyum 2022-02-03
I didn't investigate further, because dynamic import is a feature of es2020, async/await is from es2017, and we ship es2015 at this point, in umd/amd/iife/commonjs formats, so the transpilation and modularization is quite convoluted right now.

I would love for us to simplify things with a newer language standard (see #881), and maybe even esm throughout, but there are still a few blockers holding us back, such as [web workers in firefox](https://caniuse.com/mdn-api_worker_worker_ecmascript_modules), and until, I believe we might be better off avoiding too new features as they just add polyfills.